pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fpdf>=1.7.2",
    "google-genai>=1.8.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "tabulate>=0.9.0",
]
